working but ugly

+++ src/area.rs
@@ -1,9 +1,11 @@
+use crate::Square;
+
 pub trait Area {
-    fn area(&self) -> f64;
+    fn area(&self) -> i32;
 }
 
 impl Area for Square {
-    fn area(&self) -> f64 {
+    fn area(&self) -> i32 {
         self.side * self.side
     }
 }
